requirements.txt
aiohttp
PyYAML
tqdm